% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_runner_outcome_model.R
\name{estimate_runner_outcome_model}
\alias{estimate_runner_outcome_model}
\title{Estimate logistic mixed-effects models for runner outcomes}
\usage{
estimate_runner_outcome_model(
  data,
  bootstrap_index = NULL,
  bootstrap_simulate = NULL,
  bootstrap_fit_original = NULL,
  verbose = FALSE,
  include_pickoff_attempt = TRUE,
  include_pickoff_success = TRUE,
  include_runner_going = TRUE,
  include_going_interrupt = TRUE,
  include_stolen_base = TRUE
)
}
\arguments{
\item{data}{A data frame containing play-by-play baseball data. Required
columns include \code{pre_runner_1b_id}, \code{is_1b_only},
\code{is_full_count_two_outs}, \code{is_going_interrupt}, \code{runner_id},
\code{year}, \code{pre_disengagements}, \code{is_pickoff_attempt},
\code{is_pickoff_success}, \code{is_stolen_base}, \code{pre_outs},
\code{pre_balls}, \code{pre_strikes}, \code{lead_distance_centered},
\code{sprint_speed_centered}, \code{arm_strength_centered},
\code{pitcher_id}, \code{runner_id}, and \code{catcher_id}.}

\item{bootstrap_index}{integer, optional, indicating the index of the bootstrap sample
to use. If NULL (default), game model is estimated using original data.}

\item{bootstrap_simulate}{optimal list of bootstrap parametric simulations from
\code{glmmTMB:::simulate.glmmTMB}, ignored if \code{bootstrap_index} is NULL.}

\item{bootstrap_fit_original}{optional list of original fitted \code{glmmTMB} objects
for refitting, ignored if \code{bootstrap_index} is NULL.}

\item{verbose}{Logical. If \code{TRUE}, logs estimation progress to the console.
Defaults to \code{FALSE}.}

\item{include_pickoff_attempt}{logical, should pickoff attempt model be estimated?}

\item{include_pickoff_success}{logical, should pickoff success model be estimated?}

\item{include_runner_going}{logical, should runner going model be estimated?}

\item{include_going_interrupt}{logical, should batter interruption model be estimated?}

\item{include_stolen_base}{logical, should stolen base success model be estimated?}
}
\value{
A named list containing four fitted \code{glmerMod} objects:
\itemize{
\item \code{pickoff_attempt}: GLMM for the probability of a pickoff attempt.
\item \code{pickoff_success}: GLMM for the probability of a successful pickoff.
\item \code{runner_going}: GLMM for the probability of the runner going.
\item \code{going_interrupt}: GLMM for the probability of a batter interruption.
\item \code{stolen_base}: GLMM for the probability of a successful stolen base.
}
}
\description{
Fits a series of Generalized Linear Mixed Models (GLMMs) to estimate the
probabilities of pickoff (PO) attempts/success and stolen base (SB)
attempts/success. The models account for game-state context, player-specific
random effects, and environmental factors like disengagement counts.
}
