% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce.glmmTMB.R
\name{extract_fixef}
\alias{extract_fixef}
\title{Extract fixed effects from glmmTMB or glmmTMBreduced object

This is just a helper function to facilitate identifying the class of the object and then
using the appropriate code to extract the estimated fixed effects.}
\usage{
extract_fixef(object)
}
\arguments{
\item{object}{a \code{glmmTMB} or \code{glmmTMBreduced} object}
}
\value{
a named vector of fixed effects
}
\description{
Extract fixed effects from glmmTMB or glmmTMBreduced object

This is just a helper function to facilitate identifying the class of the object and then
using the appropriate code to extract the estimated fixed effects.
}
