% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_percentile_players.R
\name{extract_percentile_players}
\alias{extract_percentile_players}
\title{Extract representative players at specified percentiles}
\usage{
extract_percentile_players(
  object,
  data,
  percentile = c(0.1, 0.5, 0.9),
  flip = FALSE
)
}
\arguments{
\item{object}{A fitted \code{glmmTMB} model object containing
random effects for \code{runner_id}, \code{pitcher_id}, and/or \code{catcher_id},
and possibly fixed effects for \code{sprint_speed_centered} and \code{arm_strength_centered}.}

\item{data}{A table containing the data used to train \code{object}, including
\code{runner_id}, \code{pitcher_id}, \code{catcher_id}, \code{sprint_speed_centered},
and \code{arm_strength_centered}.}

\item{percentile}{A numeric vector of percentiles to extract. Defaults to
\code{c(0.1, 0.5, 0.9)}.}

\item{flip}{logical, should percentiles be flipped so that larger numbers are better for the
battery and smaller numbers are better for the runner? Defaults to FALSE.}
}
\value{
A list with two elements:
\itemize{
\item \code{runner}: A data frame of \code{runner_id} and \code{sprint_speed_centered}
at the specified percentiles.
\item \code{battery}: A data frame of \code{pitcher_id}, \code{catcher_id}, and
\code{arm_strength_centered} at the specified percentiles.
}
}
\description{
Identifies runner and battery combinations (pitcher/catcher) that correspond
to specific performance percentiles based on fixed effects and random effects
extracted from a fitted lme4 model.
}
