% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce.glmmTMB.R
\name{extract_ranef}
\alias{extract_ranef}
\title{Extract random effects from glmmTMB or glmmTMBreduced object}
\usage{
extract_ranef(object)
}
\arguments{
\item{object}{a \code{glmmTMB} or \code{glmmTMBreduced} object}
}
\value{
a list of random effect tables
}
\description{
This is just a helper function to facilitate identifying the class of the object and then
using the appropriate code to extract the predicted random effects.
}
