% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_runner_outcome_component.R
\name{predict_runner_outcome_component}
\alias{predict_runner_outcome_component}
\title{Predict component probability for runner outcome model}
\usage{
predict_runner_outcome_component(
  object,
  newdata,
  include_ranef = TRUE,
  allow_new_levels = TRUE
)
}
\arguments{
\item{object}{a \code{glmmTMB} or \code{glmmTMBreduced} object}

\item{newdata}{a new data table holding covariates for which to make predictions}

\item{include_ranef}{logical, should random effects be included in from predcitions?
Defaults to \code{TRUE}.}

\item{allow_new_levels}{logical, should we allow new levels for random effects? This plugs
directly into \code{allow.new.levels} for \code{glmmTMB} objects. For \code{glmmTMBreduced}
objects, the prediction is NA for new levels if \code{allow_new_levels} is \code{FALSE}.
Defaults to \code{TRUE}.}
}
\value{
a vector of probability predictions
}
\description{
We need this function because we introduced a \code{glmmTMBreduced} object type, which strips
away all but the essentials for prediction from a fitted \code{glmmTMB} object. This function
identifies the object class and then calls the predict method (if appropriate) or else manually
constucts the predictions from matrix multiplication (if using the reduced object).
}
