% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_runner_outcome_prob.R
\name{predict_runner_outcome_prob}
\alias{predict_runner_outcome_prob}
\title{Predict runner outcome probabilities}
\usage{
predict_runner_outcome_prob(
  runner_outcome_grid,
  fit_runner_outcome,
  percentile_players = NULL,
  runner_percentile = NULL,
  battery_percentile = NULL
)
}
\arguments{
\item{runner_outcome_grid}{A data frame containing game states and potential
actions. Must include columns such as \code{is_1b_only},
\code{is_full_count_two_outs}, \code{action_pitcher}, \code{pre_state},
and \code{action_runner}.}

\item{fit_runner_outcome}{A named list of fitted model objects (e.g., from
\code{lme4} or \code{glm}) corresponding to \code{po_attempt},
\code{po_success}, \code{sb_attempt}, and \code{sb_success}.}

\item{percentile_players}{A list of tables of representative players corresponding
to skill percentiles, from \code{\link{extract_percentile_players}}. Each table
in the list corresponds to one of the lme4 models for runner outcomes and has a
row of representative players for each percentile. The default, \code{NULL},
creates predictions assuming average players, without using random effects.}

\item{runner_percentile}{Numeric. The percentile value to filter for the
runner's ability when \code{percentile_players} is provided.}

\item{battery_percentile}{Numeric. The percentile value to filter for the
battery's ability when \code{percentile_players} is provided.}
}
\value{
A data frame containing the \code{pre_state}, \code{action_runner},
\code{action_pitcher}, \code{runner_outcome} (e.g., "P+", "S+"), and
the resulting probability \code{prob}.
}
\description{
Predicts stolen base and pickoff attempt and success probabilties from fitted lme4
models and converts these conditional probabilities to overall runner outcome
probabilities (successful pickoff P+, unsuccessful pickoff P-, successful steal S+,
unsuccessful steal S-, or no runner involvement N).
}
\details{
Calculates the probabilities of various runner outcomes (pickoff success/failure,
stolen base success/failure, or no action) based on a grid of game states,
a list of fitted regression models, and optional player-specific percentiles.
}
