% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce.glmmTMB.R
\name{reduce.glmmTMB}
\alias{reduce.glmmTMB}
\title{Reduce a glmmTMB object}
\usage{
reduce.glmmTMB(object)
}
\arguments{
\item{object}{a fitted \code{glmmTMB} object}
}
\value{
A named list:
\itemize{
\item \code{fixef}: a named vector of estimated fixed effects
\item \code{ranef}: a list of predicted random effects
\item \code{formula_fixef}: a formula object with the fixed-effects-only RHS
\item \code{summary}: a sanitized model summary with only coefficients and varcor
}
}
\description{
This function reduces a fitted glmmTMB model to keep only the formula, estimated fixed effects,
and predicted random effects, all that is needed to generate predictions from the model for new
data. Importantly, this does not include the original dataset on which the model was trained,
which is necessary for bootstrapping but impossible for restricted data.
}
