% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_state.R
\name{update_state}
\alias{update_state}
\title{Update game state}
\usage{
update_state(
  state,
  new_first = NA,
  new_bases = NA,
  new_outs = NA,
  new_balls = NA,
  new_strikes = NA,
  new_disengagements = NA
)
}
\arguments{
\item{state}{Character string formatted as "{bases}\emph{{outs}}{balls}{strikes}_{disengagements}"
(e.g., "100_1_21_0").}

\item{new_bases}{Optional character string to replace current base occupancy. Defaults to \code{NA}.}

\item{new_outs}{Optional integer to replace current out count. Defaults to \code{NA}.}

\item{new_balls}{Optional integer to replace current ball count. Defaults to \code{NA}.}

\item{new_strikes}{Optional integer to replace current strike count. Defaults to \code{NA}.}

\item{new_disengagements}{Optional integer to replace current disengagement count. Defaults to \code{NA}.}
}
\value{
An updated character string formatted as a standardized game state.
}
\description{
Modifies select attributes of the game state string by deconstructing
the original string, replacing specified values, and reconstructing the updated identifier.
}
\details{
Arguments left as \code{NA} retain their original values from the input
\code{state} string using \code{dplyr::coalesce}.
}
