% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_runner_outcome_model.R
\name{estimate_runner_outcome_model}
\alias{estimate_runner_outcome_model}
\title{Estimate logistic mixed-effects models for runner outcomes}
\usage{
estimate_runner_outcome_model(
  data,
  verbose = FALSE,
  include_po_attempt = TRUE,
  include_po_success = TRUE,
  include_sb_attempt = TRUE,
  include_sb_success = TRUE
)
}
\arguments{
\item{data}{A data frame containing play-by-play baseball data. Required
columns include \code{pre_runner_1b_id}, \code{is_1b_only},
\code{is_full_count_two_outs}, \code{is_sb_attempt}, \code{runner_id},
\code{year}, \code{pre_disengagements}, \code{is_po_attempt},
\code{is_po_success}, \code{is_stolen_base}, \code{pre_outs},
\code{pre_balls}, \code{pre_strikes}, \code{lead_distance_centered},
\code{sprint_speed_centered}, \code{arm_strength_centered},
\code{pitcher_id}, \code{runner_id}, and \code{catcher_id}.}

\item{verbose}{Logical. If \code{TRUE}, logs estimation progress to the console.
Defaults to \code{FALSE}.}

\item{include_po_attempt}{logical, should pickoff attempt model be estimated?}

\item{include_po_success}{logical, should pickoff success model be estimated?}

\item{include_sb_attempt}{logical, should stolen base attempt model be estimated?}

\item{include_sb_success}{logical, should stolen base success model be estimated?}
}
\value{
A named list containing four fitted \code{glmerMod} objects:
\itemize{
\item \code{po_attempt}: GLMM for the probability of a pickoff attempt.
\item \code{po_success}: GLMM for the probability of a successful pickoff.
\item \code{sb_attempt}: GLMM for the probability of a stolen base attempt.
\item \code{sb_success}: GLMM for the probability of a successful stolen base.
}
}
\description{
Fits a series of Generalized Linear Mixed Models (GLMMs) to estimate the
probabilities of pickoff (PO) attempts/success and stolen base (SB)
attempts/success. The models account for game-state context, player-specific
random effects, and environmental factors like disengagement counts.
}
